Remove modifiers feature
The main use case I had in mind for modifiers was to be able to modify the request before it's sent. For example, inserting a unique request ID header for observability or tracking. However, you can achieve the same thing now by passing a function as the default request init. I'm removing modifiers as they add complexity and I can't think of a good use case for modifying the response or error.